+++ database/migrations/2025_06_07_165451_create_person_election.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('person_election', function (Blueprint $table) {
             $table->comment('حوزه انتخابیه نماینده');
             $table->id('row_id');
-            $table->integer('person_id');
-            $table->integer('election_id');
+            $table->unsignedBigInteger('person_id');
+            $table->unsignedBigInteger('election_id');
             $table->timestamps();
             $table->softDeletes();
         });
