+++ database/migrations/2025_06_07_165600_create_person_commission.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('person_commission', function (Blueprint $table) {
             $table->comment('کمیسیون نماینده');
             $table->id('row_id');
-            $table->integer('person_id');
-            $table->integer('commission_id');
+            $table->unsignedBigInteger('person_id');
+            $table->unsignedBigInteger('commission_id');
             $table->timestamps();
             $table->softDeletes();
         });
